requirements.txt
# FastAPI and ASGI
fastapi==0.115.6
uvicorn[standard]==0.34.0
starlette==0.41.3

# Pydantic
pydantic==2.10.4
pydantic-settings==2.7.1

# Async SQLAlchemy + Postgres
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0

# JWT validation
python-jose[cryptography]==3.3.0
httpx==0.28.1

# ULID generation
python-ulid==3.0.0

# Type hints
typing_extensions==4.12.2
